+++ zwave-js/zwave-js.js
@@ -1,16 +1,15 @@
 module.exports = function (RED) {
     const SP = require("serialport");
-    const FMaps = require('./FunctionMaps.json')
     const Path = require('path')
     const ModulePackage = require('../package.json')
     const ZWaveJS = require('zwave-js')
-    const { Duration, createDefaultTransportFormat, CommandClasses } = require("@zwave-js/core");
+    const { createDefaultTransportFormat, CommandClasses } = require("@zwave-js/core");
     const ZWaveJSPackage = require('zwave-js/package.json')
     const Winston = require("winston");
 
     const UI = require('./ui/server.js')
     UI.init(RED)
-    let NodeList = [];
+    const NodeList = [];
 
     function Init(config) {
 
@@ -23,7 +22,7 @@ module.exports = function (RED) {
         var Logger;
         var FileTransport;
 
-        let NodeStats = {}
+        const NodeStats = {}
         var ControllerStats;
 
         // Log function
@@ -50,599 +49,6 @@ module.exports = function (RED) {
                 Logger.log(logEntry)
             }
         }
-
-        /* START DEPRECATION */
-
-        function ProcessDurationClass(Class, Operation, Params) {
-            if (Params.length > 1) {
-                if (typeof Params[1] === "object") {
-                    let Keys = Object.keys(Params[1]);
-                    if (Keys.length === 1 && Keys[0] === "Duration") {
-                        let D = new Duration(Params[1].Duration.value, Params[1].Duration.unit)
-                        Params[1] = D;
-                    }
-                }
-            }
-            return Params;
-        }
-
-        function ParseMeterOptions(Class, Operation, Params) {
-            if (typeof Params[0] === "object") {
-                Params[0].rateType = Enums.RateType[Params[0].rateType];
-            }
-            return Params;
-
-        }
-
-        async function OLDAssociations(msg, send) {
-
-            let Operation = msg.payload.operation
-            let Params = msg.payload.params || [];
-
-            let ReturnNode = { id: "" };
-            switch (Operation) {
-                case "GetAssociationGroups":
-                    NodeCheck(Params[0].nodeId);
-                    var ResultData = Driver.controller.getAssociationGroups(Params[0])
-                    var PL = []
-                    ResultData.forEach((FV, FK) => {
-                        let A = {
-                            GroupID: FK,
-                            AssociationGroupInfo: FV
-                        }
-                        PL.push(A);
-                    })
-
-                    ReturnNode.id = Params[0].nodeId
-                    Send(ReturnNode, "ASSOCIATION_GROUPS", { SourceAddress: Params[0], Groups: PL }, send)
-                    break;
-
-                case "GetAllAssociationGroups":
-                    NodeCheck(Params[0]);
-                    var ResultData = Driver.controller.getAllAssociationGroups(Params[0])
-                    var PL = [];
-                    ResultData.forEach((FV, FK) => {
-                        let A = {
-                            Endpoint: FK,
-                            Groups: []
-                        }
-                        FV.forEach((SV, SK) => {
-                            let B = {
-                                GroupID: SK,
-                                AssociationGroupInfo: SV
-                            }
-                            A.Groups.push(B)
-                        })
-                        PL.push(A);
-                    })
-
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "ALL_ASSOCIATION_GROUPS", PL, send)
-                    break;
-
-                case "GetAssociations":
-                    NodeCheck(Params[0].nodeId);
-                    var ResultData = Driver.controller.getAssociations(Params[0])
-                    var PL = []
-                    ResultData.forEach((FV, FK) => {
-                        let A = {
-                            GroupID: FK,
-                            AssociationAddress: []
-                        }
-                        FV.forEach((AA) => {
-                            A.AssociationAddress.push(AA);
-                        });
-
-                        PL.push(A)
-                    })
-
-                    ReturnNode.id = Params[0].nodeId
-                    Send(ReturnNode, "ASSOCIATIONS", { SourceAddress: Params[0], Associations: PL }, send)
-                    break;
-
-                case "GetAllAssociations":
-                    NodeCheck(Params[0]);
-                    var ResultData = Driver.controller.getAllAssociations(Params[0]);
-                    var PL = []
-                    ResultData.forEach((FV, FK) => {
-                        let A = {
-                            AssociationAddress: FK,
-                            Associations: []
-                        }
-                        FV.forEach((SV, SK) => {
-                            let B = {
-                                GroupID: SK,
-                                AssociationAddress: SV
-                            }
-                            A.Associations.push(B)
-                        });
-                        PL.push(A)
-                    })
-
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "ALL_ASSOCIATIONS", PL, send)
-                    break;
-
-                case "AddAssociations":
-                    NodeCheck(Params[0].nodeId);
-                    Params[2].forEach((A) => {
-                        if (!Driver.controller.isAssociationAllowed(Params[0], Params[1], A)) {
-                            let ErrorMSG = "Association: Source " + JSON.stringify(Params[0]); +", Group " + Params[1] + ", Destination " + SON.stringify(A) + " is not allowed."
-                            throw new Error(ErrorMSG);
-                        }
-                    })
-                    await Driver.controller.addAssociations(Params[0], Params[1], Params[2])
-                    ReturnNode.id = Params[0].nodeId
-                    Send(ReturnNode, "ASSOCIATIONS_ADDED", undefined, send)
-                    break;
-
-                case "RemoveAssociations":
-                    NodeCheck(Params[0].nodeId);
-                    await Driver.controller.removeAssociations(Params[0], Params[1], Params[2])
-                    ReturnNode.id = Params[0].nodeId
-                    Send(ReturnNode, "ASSOCIATIONS_REMOVED", undefined, send)
-                    break;
-
-                case "RemoveNodeFromAllAssociations":
-                    NodeCheck(Params[0]);
-                    await Driver.controller.removeNodeFromAllAssociations(Params[0])
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "ALL_ASSOCIATIONS_REMOVED", undefined, send)
-                    break;
-            }
-
-            return;
-        }
-
-
-        async function OLDController(msg, send) {
-
-            let Operation = msg.payload.operation
-            let Params = msg.payload.params || [];
-
-            let ReturnController = { id: "Controller" };
-            let ReturnNode = { id: "" };
-
-            let SupportsNN = false;
-
-            switch (Operation) {
-
-                case "AbortFirmwareUpdate":
-                    NodeCheck(Params[0])
-                    ReturnNode.id = Params[0]
-                    await Driver.controller.nodes.get(Params[0]).abortFirmwareUpdate()
-                    Send(ReturnNode, "FIRMWARE_UPDATE_ABORTED", undefined, send);
-                    break;
-
-                case "BeginUpdateFirmware":
-                    NodeCheck(Params[0])
-                    ReturnNode.id = Params[0]
-                    let Format = ZWaveJS.guessFirmwareFileFormat(Params[2], Params[3])
-                    let Firmware = ZWaveJS.extractFirmware(Params[3], Format)
-                    await Driver.controller.nodes.get(Params[0]).beginFirmwareUpdate(Firmware.data, Params[1])
-                    Send(ReturnNode, "FIRMWARE_UPDATE_STARTED", Params[1], send);
-                    break;
-
-                case "GetRFRegion":
-                    let RFR = await Driver.controller.getRFRegion();
-                    Send(ReturnController, "CURRENT_RF_REGION", Enums.RFRegion[RFR], send);
-                    break;
-
-                case "SetRFRegion":
-                    await Driver.controller.setRFRegion(Enums.RFRegion[Params[0]]);
-                    Send(ReturnController, "RF_REGION_SET", Params[0], send);
-                    break;
-
-                case "ToggleRF":
-                    await Driver.controller.toggleRF(Params[0]);
-                    Send(ReturnController, "RF_STATUS", Params[0], send);
-                    break;
-
-                case "GetNodes":
-                    let Nodes = [];
-                    Driver.controller.nodes.forEach((N, NI) => {
-                        Nodes.push({
-                            nodeId: N.id,
-                            name: N.name,
-                            location: N.location,
-                            status: Enums.NodeStatus[N.status],
-                            ready: N.ready,
-                            interviewStage: Enums.InterviewStage[N.interviewStage],
-                            zwavePlusVersion: N.zwavePlusVersion,
-                            zwavePlusNodeType: N.zwavePlusNodeType,
-                            zwavePlusRoleType: N.zwavePlusRoleType,
-                            isListening: N.isListening,
-                            isFrequentListening: N.isFrequentListening,
-                            canSleep: N.canSleep,
-                            isRouting: N.isRouting,
-                            supportedDataRates: N.supportedDataRates,
-                            maxDataRate: N.maxDataRate,
-                            supportsSecurity: N.supportsSecurity,
-                            isSecure: N.isSecure,
-                            protocolVersion: Enums.ProtocolVersion[N.protocolVersion],
-                            manufacturerId: N.manufacturerId,
-                            productId: N.productId,
-                            productType: N.productType,
-                            firmwareVersion: N.firmwareVersion,
-                            deviceConfig: N.deviceConfig,
-                            isControllerNode: N.isControllerNode(),
-                            supportsBeaming: N.supportsBeaming,
-                            keepAwake: N.keepAwake
-                        })
-                    });
-                    Send(ReturnController, "NODE_LIST", Nodes, send);
-                    break;
-
-                case "KeepNodeAwake":
-                    NodeCheck(Params[0])
-                    ReturnNode.id = Params[0]
-                    Driver.controller.nodes.get(Params[0]).keepAwake = Params[1]
-                    Send(ReturnNode, "NODE_KEEP_AWAKE", Params[1], send)
-                    break;
-
-                case "GetNodeNeighbors":
-                    NodeCheck(Params[0])
-                    let NIDs = await Driver.controller.getNodeNeighbors(Params[0]);
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "NODE_NEIGHBORS", NIDs, send);
-                    break;
-
-                case "SetNodeName":
-                    NodeCheck(Params[0])
-                    Driver.controller.nodes.get(Params[0]).name = Params[1]
-                    SupportsNN = Driver.controller.nodes.get(Params[0]).supportsCC(CommandClasses["Node Naming and Location"])
-                    if (SupportsNN) {
-                        await Driver.controller.nodes.get(Params[0]).commandClasses["Node Naming and Location"].setName(Params[1]);
-                    }
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "NODE_NAME_SET", Params[1], send)
-                    ShareNodeList();
-                    break
-
-                case "SetNodeLocation":
-                    NodeCheck(Params[0])
-                    Driver.controller.nodes.get(Params[0]).location = Params[1]
-                    SupportsNN = Driver.controller.nodes.get(Params[0]).supportsCC(CommandClasses["Node Naming and Location"])
-                    if (SupportsNN) {
-                        await Driver.controller.nodes.get(Params[0]).commandClasses["Node Naming and Location"].setLocation(Params[1]);
-                    }
-                    ReturnNode.id = Params[0]
-                    Send(ReturnNode, "NODE_LOCATION_SET", Params[1], send)
-                    break
-
-                case "InterviewNode":
-                    Params[0] = +Params[0]
-                    NodeCheck(Params[0], true);
-                    let Stage = Enums.InterviewStage[Driver.controller.nodes.get(Params[0]).interviewStage];
-                    if (Stage !== "Complete") {
-                        let ErrorMSG = "Node " + Params[0] + " is already being interviewed. Current Interview Stage : " + Stage + "";
-                        throw new Error(ErrorMSG);
-                    }
-                    else {
-                        await Driver.controller.nodes.get(Params[0]).refreshInfo();
-                    }
-                    break;
-
-                case "HardReset":
-                    await Driver.hardReset();
-                    Send(ReturnController, "CONTROLLER_RESET_COMPLETE", undefined, send)
-                    break;
-
-                case "StartHealNetwork":
-                    await Driver.controller.beginHealingNetwork();
-                    Send(ReturnController, "NETWORK_HEAL_STARTED", undefined, send)
-                    node.status({ fill: "yellow", shape: "dot", text: "Network Heal Started." });
-                    UI.status("Network Heal Started.")
-                    break;
-
-                case "StopHealNetwork":
-                    await Driver.controller.stopHealingNetwork();
-                    Send(ReturnController, "NETWORK_HEAL_STOPPED", undefined, send)
-                    node.status({ fill: "blue", shape: "dot", text: "Network Heal Stopped." });
-                    UI.status("Network Heal Stopped.")
-                    RestoreReadyStatus();
-                    break;
-
-                case "RemoveFailedNode":
-                    await Driver.controller.removeFailedNode(Params[0]);
-                    break;
-
-                case "ReplaceFailedNode":
-                    if (!canDoSecure) {
-                        await Driver.controller.replaceFailedNode(Params[0], true);
-                    }
-                    else if (Params.length > 1) {
-                        await Driver.controller.replaceFailedNode(Params[0], Params[1]);
-                    }
-                    else {
-                        await Driver.controller.replaceFailedNode(Params[0], false);
-                    }
-                    break;
-
-                case "StartInclusion":
-                    if (!canDoSecure) {
-                        await Driver.controller.beginInclusion(true);
-                    }
-                    else if (Params !== undefined && Params.length > 0) {
-                        await Driver.controller.beginInclusion(Params[0]);
-                    }
-                    else {
-                        await Driver.controller.beginInclusion(false);
-                    }
-                    break;
-
-                case "StopInclusion":
-                    await Driver.controller.stopInclusion();
-                    break;
-
-                case "StartExclusion":
-                    await Driver.controller.beginExclusion();
-                    break;
-
-                case "StopExclusion":
-                    await Driver.controller.stopExclusion();
-                    break;
-
-                case "ProprietaryFunc":
-
-                    let ZWaveMessage = new ZWaveJS.Message(Driver, {
-                        type: ZWaveJS.MessageType.Request,
-                        functionType: Params[0],
-                        payload: Params[1]
-                    })
-
-
-                    let MessageSettings = {
-                        priority: ZWaveJS.MessagePriority.Controller,
-                        supportCheck: false
-                    }
-
-                    await Driver.sendMessage(ZWaveMessage, MessageSettings)
-                    break;
-
-
-            }
-
-            return;
-        }
-
-        async function OLDUnmanaged(msg, send) {
-
-            let Operation = msg.payload.operation
-            let Node = msg.payload.node;
-            let Params = msg.payload.params || [];
-
-            let ReturnNode = { id: Node };
-
-            NodeCheck(Node);
-
-            switch (Operation) {
-                case "GetDefinedValueIDs":
-                    const VIDs = Driver.controller.nodes.get(Node).getDefinedValueIDs();
-                    Send(ReturnNode, "VALUE_ID_LIST", VIDs, send);
-                    break;
-
-                case "SetValue":
-                    await Driver.controller.nodes.get(Node).setValue(Params[0], Params[1]);
-                    break;
-
-                case "GetValue":
-                    let V = Driver.controller.nodes.get(Node).getValue(Params[0]);
-
-                    let ReturnObject = {
-                        response: V,
-                        valueId: Params[0]
-                    }
-                    Send(ReturnNode, "GET_VALUE_RESPONSE", ReturnObject, send);
-                    break;
-
-                case "GetValueMetadata":
-                    let M = Driver.controller.nodes.get(Node).getValueMetadata(Params[0]);
-
-                    let ReturnObjectM = {
-                        response: M,
-                        valueId: Params[0]
-                    }
-                    Send(ReturnNode, "GET_VALUE_METADATA_RESPONSE", ReturnObjectM, send);
-                    break;
-
-                case "PollValue":
-                    await Driver.controller.nodes.get(Node).pollValue(Params[0]);
-                    break;
-
-            }
-
-            return;
-        }
-
-        // Z-Wave JS Enum Lookups
-        const Enums = {
-            // CC enums
-            RateType: ZWaveJS.RateType,
-            ColorComponent: ZWaveJS.ColorComponent,
-            SetbackType: ZWaveJS.SetbackType,
-            BinarySensorType: ZWaveJS.BinarySensorType,
-            ThermostatMode: ZWaveJS.ThermostatMode,
-            SetPointType: ZWaveJS.ThermostatSetpointType,
-            DoorLockMode: ZWaveJS.DoorLockMode,
-            AlarmSensorType: ZWaveJS.AlarmSensorType,
-            BarrierState: ZWaveJS.BarrierState,
-            SubsystemType: ZWaveJS.SubsystemType,
-            SubsystemState: ZWaveJS.SubsystemState,
-            UserIDStatus: ZWaveJS.UserIDStatus,
-            KeypadMode: ZWaveJS.KeypadMode,
-            Weekday: ZWaveJS.Weekday,
-
-            // Controller Enums
-            RFRegion: ZWaveJS.RFRegion,
-
-            // node enums
-            InterviewStage: ZWaveJS.InterviewStage,
-            NodeStatus: ZWaveJS.NodeStatus,
-            ProtocolVersion: ZWaveJS.ProtocolVersion
-        }
-
-        const CCParamConverters = {
-            "BinarySwitch.Set": ProcessDurationClass,
-            "MultiLevelSwitch.Set": ProcessDurationClass,
-            "Meter.Get": ParseMeterOptions
-        }
-
-
-        async function OldInput(msg, send) {
-
-
-            let Class = msg.payload.class;
-
-            switch (Class) {
-                case "Controller":
-                    await OLDController(msg, send)
-                    break;
-
-                case "Unmanaged":
-                    await OLDUnmanaged(msg, send);
-                    break;
-
-                case "Driver":
-                    await OLDDriverCMD(msg, send);
-                    break;
-
-                case "Associations":
-                    await OLDAssociations(msg, send);
-                    break;
-
-                default:
-                    await OLDNodeFunction(msg, send);
-                    break;
-            }
-
-            return;
-
-        }
-
-        async function OLDNodeFunction(msg, send) {
-
-            let Operation = msg.payload.operation
-            let Class = msg.payload.class;
-            let Node = msg.payload.node
-            var Params = msg.payload.params || [];
-            let forceUpdate = msg.payload.forceUpdate;
-
-            let ReturnNode = { id: Node };
-
-            NodeCheck(Node);
-
-            if (!FMaps.hasOwnProperty(Class)) {
-                let ErrorMSG = "Class, " + Class + " not supported.";
-                throw new Error(ErrorMSG);
-            }
-
-            let Map = FMaps[Class];
-
-            if (!Map.Operations.hasOwnProperty(Operation)) {
-                let ErrorMSG = "Unsupported operation : " + Operation + " for class " + Class;
-                throw new Error(ErrorMSG);
-            }
-
-            let Func = Map.Operations[Operation];
-
-            if (Params.length !== Func.ParamsRequired && Params.length !== (Func.ParamsOptional + Func.ParamsRequired)) {
-                let ErrorMSG = "Incorrect number of parameters specified for " + Operation;
-                throw new Error(ErrorMSG);
-            }
-
-            let EP = 0;
-
-            if (msg.payload.hasOwnProperty("endpoint")) {
-                EP = parseInt(msg.payload.endpoint)
-            }
-
-            if (Func.hasOwnProperty("ParamEnumDependency")) {
-                for (let i = 0; i < Params.length; i++) {
-                    if (Func.ParamEnumDependency.hasOwnProperty(i.toString())) {
-                        let Enum = Func.ParamEnumDependency[i.toString()];
-                        Params[i] = Enums[Enum][Params[i]]
-                    }
-                }
-            }
-
-            if (CCParamConverters.hasOwnProperty(Class + "." + Operation)) {
-                let Handler = CCParamConverters[Class + "." + Operation];
-                Params = Handler(Class, Operation, Params);
-            }
-
-            let ZWJSC = Driver.controller.nodes.get(Node).getEndpoint(EP).commandClasses[Map.MapsToClass];
-
-            let WaitForResponse = (Func.hasOwnProperty("NoEvent") && Func.NoEvent);
-
-            if (WaitForResponse) {
-                let Result = await ZWJSC[Func.MapsToFunc].apply(ZWJSC, Params);
-                Send(ReturnNode, "VALUE_UPDATED", Result, send)
-            }
-            else {
-                await ZWJSC[Func.MapsToFunc].apply(ZWJSC, Params);
-            }
-
-            if (forceUpdate !== undefined) {
-                let VID = {
-                    commandClass: CommandClasses[Map.MapsToClass],
-                    endpoint: EP
-                }
-                Object.keys(forceUpdate).forEach((VIDK) => {
-                    VID[VIDK] = forceUpdate[VIDK]
-                })
-                await Driver.controller.nodes.get(Node).pollValue(VID)
-            }
-
-            return;
-        }
-
-        async function OLDDriverCMD(msg, send) {
-
-            let Operation = msg.payload.operation;
-            let Params = msg.payload.params || []
-            let ReturnNode = { id: "N/A" };
-
-            switch (Operation) {
-
-                case "GetEnums":
-                    Send(ReturnNode, "ENUM_LIST", Enums, send);
-                    break;
-
-                case "GetValueDB":
-
-                    let Result = [];
-
-                    if (Params.length < 1) {
-                        Driver.controller.nodes.forEach((N, NI) => {
-                            Params.push(N.id)
-                        });
-                    }
-                    Params.forEach((NID) => {
-                        let G = {
-                            nodeId: NID,
-                            nodeName: getNodeInfoForPayload(NID, 'name'),
-                            nodeLocation: getNodeInfoForPayload(NID, 'location'),
-                            values: []
-                        }
-                        const VIDs = Driver.controller.nodes.get(NID).getDefinedValueIDs();
-                        VIDs.forEach((VID) => {
-                            let V = Driver.controller.nodes.get(NID).getValue(VID);
-                            let VI = {
-                                currentValue: V,
-                                valueId: VID
-                            }
-                            G.values.push(VI)
-                        })
-                        Result.push(G);
-                    })
-                    Send(ReturnNode, "VALUE_DB", Result, send);
-                    break;
-            }
-        }
-
-
-        /* END DEPRECATION */
 
         var RestoreReadyTimer;
         function RestoreReadyStatus() {
@@ -684,7 +90,7 @@ module.exports = function (RED) {
             await Input(MSG, undefined, undefined, true)
         }
 
-        let DriverOptions = {};
+        const DriverOptions = {};
 
         // Logging
         DriverOptions.logConfig = {};
@@ -712,7 +118,7 @@ module.exports = function (RED) {
         // Cache Dir
         Log("debug", "NDERED", undefined, "[options] [storage.cacheDir]", Path.join(RED.settings.userDir, "zwave-js-cache"))
         DriverOptions.storage.cacheDir = Path.join(RED.settings.userDir, "zwave-js-cache");
-        
+
 
         // Custom  Config Path
         if (config.customConfigPath !== undefined && config.customConfigPath.length > 0) {
@@ -746,7 +152,7 @@ module.exports = function (RED) {
             let RemoveBrackets = config.encryptionKey.replace("[", "").replace("]", "");
             let _Array = RemoveBrackets.split(",");
 
-            Log("debug", "NDERED", undefined, "[options] [networkKey]", "Provided as array","["+_Array+" bytes]")
+            Log("debug", "NDERED", undefined, "[options] [networkKey]", "Provided as array", "[" + _Array + " bytes]")
 
             let _Buffer = [];
             for (let i = 0; i < _Array.length; i++) {
@@ -758,7 +164,7 @@ module.exports = function (RED) {
         }
         else if (config.encryptionKey !== undefined && config.encryptionKey.length > 0) {
 
-            Log("debug", "NDERED", undefined, "[options] [networkKey]", "Provided as string","["+config.encryptionKey.length+" characters]")
+            Log("debug", "NDERED", undefined, "[options] [networkKey]", "Provided as string", "[" + config.encryptionKey.length + " characters]")
             DriverOptions.networkKey = Buffer.from(config.encryptionKey);
             canDoSecure = true;
         }
@@ -914,10 +320,10 @@ module.exports = function (RED) {
 
         });
 
-        node.on('close', (removed,done) => {
+        node.on('close', (removed, done) => {
 
             let Type = (removed ? "DELETE" : "RESTART")
-            Log("info", "NDERED", undefined, "[SHUTDOWN] ["+Type+"]", "Cleaning up...")
+            Log("info", "NDERED", undefined, "[SHUTDOWN] [" + Type + "]", "Cleaning up...")
             UI.unregister()
             Driver.destroy();
             RED.events.off("zwjs:node:command", processMessageEvent);
@@ -1034,38 +440,33 @@ module.exports = function (RED) {
         async function Input(msg, send, done, internal) {
 
             var Type = "CONTROLLER";
-            if(internal !== undefined && internal){
+            if (internal !== undefined && internal) {
                 Type = "EVENT"
             }
 
-            Log("debug", "NDERED", "IN", "["+Type+"]", "Payload received.")
+            Log("debug", "NDERED", "IN", "[" + Type + "]", "Payload received.")
 
             try {
 
-                if (msg.payload.mode === undefined) {
-                    await OldInput(msg, send)
+                let Mode = msg.payload.mode;
+                switch (Mode) {
+                    case "CCAPI":
+                        await CCAPI(msg, send);
+                        break;
+                    case "ValueAPI":
+                        await ValueAPI(msg, send)
+                        break;
+                    case "DriverAPI":
+                        await DriverAPI(msg, send)
+                        break;
+                    case "ControllerAPI":
+                        await ControllerAPI(msg, send)
+                        break;
+                    case "AssociationsAPI":
+                        await AssociationsAPI(msg, send);
+                        break;
                 }
-                else {
 
-                    let Mode = msg.payload.mode;
-                    switch (Mode) {
-                        case "CCAPI":
-                            await CCAPI(msg, send);
-                            break;
-                        case "ValueAPI":
-                            await ValueAPI(msg, send)
-                            break;
-                        case "DriverAPI":
-                            await DriverAPI(msg, send)
-                            break;
-                        case "ControllerAPI":
-                            await ControllerAPI(msg, send)
-                            break;
-                        case "AssociationsAPI":
-                            await AssociationsAPI(msg, send);
-                            break;
-                    }
-                }
 
                 if (done) {
                     done()
@@ -1301,10 +702,10 @@ module.exports = function (RED) {
             let Node = msg.payload.node;
 
             var ZWaveNode;
-            if(Array.isArray(Node)){
-                ZWaveNode =  Driver.controller.getMulticastGroup(Node)
+            if (Array.isArray(Node)) {
+                ZWaveNode = Driver.controller.getMulticastGroup(Node)
             }
-            else{
+            else {
                 NodeCheck(Node);
                 ZWaveNode = Driver.controller.nodes.get(Node)
             }
@@ -1365,11 +766,11 @@ module.exports = function (RED) {
             let EnumSelection = msg.payload.enums;
             let ForceUpdate = msg.payload.forceUpdate
 
-           
-            if(Array.isArray(Node)){
-                ZWaveNode =  Driver.controller.getMulticastGroup(Node)
+
+            if (Array.isArray(Node)) {
+                ZWaveNode = Driver.controller.getMulticastGroup(Node)
             }
-            else{
+            else {
                 NodeCheck(Node);
                 ZWaveNode = Driver.controller.nodes.get(Node)
             }
@@ -1406,7 +807,7 @@ module.exports = function (RED) {
                 Object.keys(ForceUpdate).forEach((VIDK) => {
                     ValueID[VIDK] = ForceUpdate[VIDK]
                 })
-                Log("debug", "NDERED", undefined, "[POLL]", printForceUpdate(Node,ValueID))
+                Log("debug", "NDERED", undefined, "[POLL]", printForceUpdate(Node, ValueID))
                 await ZWaveNode.pollValue(ValueID)
 
             }
@@ -1681,14 +1082,14 @@ module.exports = function (RED) {
             }
 
             let _Subject = ""
-            if(Node.id !== 'N/A' && Node.id !== 'Controller' ){
-                _Subject = "[Node: "+Node.id+"] ["+Subject+"]"
+            if (Node.id !== 'N/A' && Node.id !== 'Controller') {
+                _Subject = "[Node: " + Node.id + "] [" + Subject + "]"
             }
-            else{
-                _Subject = "["+Subject+"]"
+            else {
+                _Subject = "[" + Subject + "]"
             }
 
-            Log("debug","NDERED","OUT",_Subject,"Forwarding payload...")
+            Log("debug", "NDERED", "OUT", _Subject, "Forwarding payload...")
 
             if (send) {
                 send({ "payload": PL })
@@ -1711,11 +1112,9 @@ module.exports = function (RED) {
             if (AllowedSubjectsForDNs.includes(Subject)) {
                 RED.events.emit("zwjs:node:event:all", { "payload": PL })
                 RED.events.emit("zwjs:node:event:" + Node.id, { "payload": PL })
-                
+
             }
         }
-
-
 
         Log("info", "NDERED", undefined, undefined, "Starting Driver...")
         Driver.start()
@@ -1728,14 +1127,17 @@ module.exports = function (RED) {
 
     RED.nodes.registerType("zwave-js", Init);
 
+    /* Node List */
     RED.httpAdmin.get("/zwjsgetnodelist", function (req, res) {
         res.json(NodeList)
     })
 
+    /* Versions */
     RED.httpAdmin.get("/zwjsgetversion", function (req, res) {
         res.json({ "zwjsversion": ZWaveJSPackage.version, "moduleversion": ModulePackage.version })
     })
 
+    /* Ports */
     RED.httpAdmin.get("/zwjsgetports", RED.auth.needsPermission('serial.read'), function (req, res) {
         SP.list()
             .then(ports => {